+++ CHANGELOG.md
@@ -35,7 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `require` casuing an error when requiring a module that returns nil.
 - Fixed `websockets` yielding forever.
 
-### `0.4.0` - September 17, 2024
+## `0.4.0` - September 17, 2024
 
 ### Added
 - Added `watch`, `openFile`, and `createFile` to `@zcore/fs`. [More Info](https://scythe-technology.github.io/zune-docs/docs/api/fs)
@@ -169,7 +169,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       print("Received SIGINT")
     end)
     ```
-
 
 ### Changed
 - Updated `luau` to `0.642`.